+++ py-polars/polars/io/csv/__init__.py
@@ -1,0 +1,7 @@
+from polars.io.csv.functions import read_csv, read_csv_batched, scan_csv
+
+__all__ = [
+    "read_csv",
+    "read_csv_batched",
+    "scan_csv",
+]
