+++ py-polars/polars/io/csv/batched_reader.py
